Feat: Completed Profit and Loss Report
