+++ accounts/accounts/report/profit_and_loss/profit_and_loss.json
@@ -1,5 +1,5 @@
 {
- "add_total_row": 1,
+ "add_total_row": 0,
  "creation": "2020-08-31 20:07:48.969811",
  "disable_prepared_report": 0,
  "disabled": 0,
@@ -8,7 +8,7 @@
  "idx": 0,
  "is_standard": "Yes",
  "letter_head": "Hello",
- "modified": "2020-08-31 20:08:05.098925",
+ "modified": "2020-09-01 18:09:39.861705",
  "modified_by": "Administrator",
  "module": "Accounts",
  "name": "Profit and Loss",
